[OM-65848] RB43229 Change SMA to support float-pointing number of coupons for templates

+++ com.vmturbo.market.component/src/main/java/com/vmturbo/market/cloudscaling/sma/entities/SMAInput.java
@@ -598,7 +598,7 @@ public class SMAInput {
         }
         ComputeTierInfo computeTierInfo = computeTier.getTypeSpecificInfo().getComputeTier();
         String family = computeTierInfo.getFamily();
-        int coupons = computeTierInfo.getNumCoupons();
+        float coupons = computeTierInfo.getNumCoupons();
 
         /*
          * Find all the contexts in which this template belongs.  Iterate threw osTypes and lookup
